requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
transformers>=4.30.0
torch>=2.0.0
matplotlib>=3.4.2
jupyter>=1.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
seaborn>=0.11.0